Change 'FirstBackoffRetryInSeconds' type to double
This change allows 'FirstBackoffRetryInSeconds' property in the 'RefitGeneratorSettings' to accept double values, rather than integers only. This is done to provide more precision for backoff retry times.

+++ src/Refitter.Core/RefitGeneratorSettings.cs
@@ -246,5 +246,5 @@ public class DependencyInjectionSettings
     /// </summary>
     [JsonPropertyName("firstBackoffRetryInSeconds")]
     [JsonProperty("firstBackoffRetryInSeconds")]
-    public int FirstBackoffRetryInSeconds { get; set; } = 1;
+    public double FirstBackoffRetryInSeconds { get; set; } = 1.0;
 }
